+++ src/Models/Weather/Coordiantes.cs
@@ -5,8 +5,8 @@ namespace Models.Weather
     public class Coordiantes
     {
         [JsonProperty("lat")]
-        public long Latitude { get; set; }
+        public double Latitude { get; set; }
         [JsonProperty("lon")]
-        public long Longitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
