+++ src/Models/Weather/Wind.cs
@@ -5,7 +5,7 @@ namespace Models.Weather
     public class Wind
     {
         [JsonProperty("speed")]
-        public long Speed { get; set; }
+        public double Speed { get; set; }
         [JsonProperty("deg")]
         public int Direction { get; set; }
     }
